Number of precursors is `unsigned int`

+++ framework/materials/multi_group_xs/xsfile.cc
@@ -373,7 +373,7 @@ XSFile::Read()
                    ls,
                    ln);
 
-        for (size_t j = 0; j < num_precursors_; ++j)
+        for (unsigned int j = 0; j < num_precursors_; ++j)
         {
           OpenSnLogicalErrorIf(not HasNonZero(emission_spectra_[j]),
                                "Delayed emission spectrum for precursor " + std::to_string(j) +
